+++ services/horizon/internal/actions_path_test.go
@@ -12,9 +12,9 @@ import (
 	"github.com/go-chi/chi"
 	"github.com/stellar/go/exp/orderbook"
 	"github.com/stellar/go/protocols/horizon"
+	"github.com/stellar/go/services/horizon/internal/actions"
 	"github.com/stellar/go/services/horizon/internal/db2"
 	"github.com/stellar/go/services/horizon/internal/db2/core"
-	"github.com/stellar/go/services/horizon/internal/paths"
 	horizonProblem "github.com/stellar/go/services/horizon/internal/render/problem"
 	"github.com/stellar/go/services/horizon/internal/simplepath"
 	"github.com/stellar/go/services/horizon/internal/test"
@@ -22,16 +22,48 @@ import (
 	"github.com/stellar/go/xdr"
 )
 
-func pathFindingClient(tt *test.T, pathFinder paths.Finder, maxAssetsParamLength int) test.RequestHelper {
+func inMemoryPathFindingClient(
+	tt *test.T,
+	graph *orderbook.OrderBookGraph,
+	maxAssetsParamLength int,
+) test.RequestHelper {
 	router := chi.NewRouter()
 	findPaths := FindPathsHandler{
-		pathFinder:           pathFinder,
+		pathFinder:           simplepath.NewInMemoryFinder(graph),
 		maxAssetsParamLength: maxAssetsParamLength,
+		setLastLedgerHeader:  true,
 		coreQ:                &core.Q{tt.CoreSession()},
 	}
 	findFixedPaths := FindFixedPathsHandler{
-		pathFinder:           pathFinder,
+		pathFinder:           simplepath.NewInMemoryFinder(graph),
 		maxAssetsParamLength: maxAssetsParamLength,
+		setLastLedgerHeader:  true,
+		coreQ:                &core.Q{tt.CoreSession()},
+	}
+
+	installPathFindingRoutes(findPaths, findFixedPaths, router, false)
+	return test.NewRequestHelper(router)
+}
+
+func dbPathFindingClient(
+	tt *test.T,
+	maxAssetsParamLength int,
+) test.RequestHelper {
+	router := chi.NewRouter()
+	findPaths := FindPathsHandler{
+		pathFinder: &simplepath.Finder{
+			Q: &core.Q{tt.CoreSession()},
+		},
+		maxAssetsParamLength: maxAssetsParamLength,
+		setLastLedgerHeader:  false,
+		coreQ:                &core.Q{tt.CoreSession()},
+	}
+	findFixedPaths := FindFixedPathsHandler{
+		pathFinder: &simplepath.Finder{
+			Q: &core.Q{tt.CoreSession()},
+		},
+		maxAssetsParamLength: maxAssetsParamLength,
+		setLastLedgerHeader:  false,
 		coreQ:                &core.Q{tt.CoreSession()},
 	}
 
@@ -43,11 +75,8 @@ func TestPathActions_Index(t *testing.T) {
 	tt := test.Start(t).Scenario("paths")
 	assertions := &Assertions{tt.Assert}
 	defer tt.Finish()
-	rh := pathFindingClient(
+	rh := dbPathFindingClient(
 		tt,
-		&simplepath.Finder{
-			Q: &core.Q{tt.CoreSession()},
-		},
 		3,
 	)
 
@@ -78,6 +107,7 @@ func TestPathActions_Index(t *testing.T) {
 		w = rh.Get(uri + "?" + q.Encode())
 		assertions.Equal(200, w.Code)
 		assertions.PageOf(3, w.Body)
+		assertions.Equal("", w.Header().Get(actions.LastLedgerHeaderName))
 	}
 }
 
@@ -85,9 +115,9 @@ func TestPathActionsStillIngesting(t *testing.T) {
 	tt := test.Start(t).Scenario("paths")
 	defer tt.Finish()
 	assertions := &Assertions{tt.Assert}
-	rh := pathFindingClient(
+	rh := inMemoryPathFindingClient(
 		tt,
-		simplepath.NewInMemoryFinder(orderbook.NewOrderBookGraph()),
+		orderbook.NewOrderBookGraph(),
 		3,
 	)
 
@@ -113,6 +143,7 @@ func TestPathActionsStillIngesting(t *testing.T) {
 		w := rh.Get(uri + "?" + q.Encode())
 		assertions.Equal(horizonProblem.StillIngesting.Status, w.Code)
 		assertions.Problem(w.Body, horizonProblem.StillIngesting)
+		assertions.Equal("", w.Header().Get(actions.LastLedgerHeaderName))
 	}
 }
 
@@ -134,6 +165,7 @@ func TestPathActionsStateInvalid(t *testing.T) {
 	w := rh.Get("/paths")
 	// Still ingesting
 	rh.Assert.Equal(503, w.Code)
+	rh.Assert.Equal("", w.Header().Get(actions.LastLedgerHeaderName))
 
 	err := rh.App.historyQ.UpdateLastLedgerExpIngest(10)
 	rh.Assert.NoError(err)
@@ -144,9 +176,15 @@ func TestPathActionsStateInvalid(t *testing.T) {
 	w = rh.Get("/paths")
 	// State invalid
 	rh.Assert.Equal(500, w.Code)
+	rh.Assert.Equal("", w.Header().Get(actions.LastLedgerHeaderName))
 }
 
-func loadOffers(tt *test.T, orderBookGraph *orderbook.OrderBookGraph, fromAddress string) {
+func loadOffers(
+	tt *test.T,
+	orderBookGraph *orderbook.OrderBookGraph,
+	fromAddress string,
+	ledger uint32,
+) {
 	coreQ := &core.Q{tt.CoreSession()}
 	offers := []core.Offer{}
 	pageQuery := db2.PageQuery{
@@ -166,7 +204,7 @@ func loadOffers(tt *test.T, orderBookGraph *orderbook.OrderBookGraph, fromAddres
 			Price:    xdr.Price{N: xdr.Int32(offer.Price * 100), D: 100},
 		})
 	}
-	tt.Assert.NoError(orderBookGraph.Apply())
+	tt.Assert.NoError(orderBookGraph.Apply(ledger))
 }
 
 func TestPathActionsInMemoryFinder(t *testing.T) {
@@ -179,21 +217,18 @@ func TestPathActionsInMemoryFinder(t *testing.T) {
 	sourceAssets, _, err := coreQ.AssetsForAddress(sourceAccount)
 	tt.Assert.NoError(err)
 
-	inMemoryPathsClient := pathFindingClient(
+	inMemoryPathsClient := inMemoryPathFindingClient(
 		tt,
-		simplepath.NewInMemoryFinder(orderBookGraph),
+		orderBookGraph,
 		len(sourceAssets),
 	)
-	dbPathsClient := pathFindingClient(
+	dbPathsClient := dbPathFindingClient(
 		tt,
-		&simplepath.Finder{
-			Q: &core.Q{tt.CoreSession()},
-		},
 		len(sourceAssets),
 	)
 
-	loadOffers(tt, orderBookGraph, "GA2NC4ZOXMXLVQAQQ5IQKJX47M3PKBQV2N5UV5Z4OXLQJ3CKMBA2O2YL")
-	loadOffers(tt, orderBookGraph, "GDSBCQO34HWPGUGQSP3QBFEXVTSR2PW46UIGTHVWGWJGQKH3AFNHXHXN")
+	loadOffers(tt, orderBookGraph, "GA2NC4ZOXMXLVQAQQ5IQKJX47M3PKBQV2N5UV5Z4OXLQJ3CKMBA2O2YL", 1)
+	loadOffers(tt, orderBookGraph, "GDSBCQO34HWPGUGQSP3QBFEXVTSR2PW46UIGTHVWGWJGQKH3AFNHXHXN", 2)
 
 	var withSourceAccount = make(url.Values)
 	withSourceAccount.Add(
@@ -224,11 +259,13 @@ func TestPathActionsInMemoryFinder(t *testing.T) {
 		tt.Assert.Equal(http.StatusOK, w.Code)
 		inMemorySourceAccountResponse := []horizon.Path{}
 		tt.UnmarshalPage(w.Body, &inMemorySourceAccountResponse)
+		tt.Assert.Equal("2", w.Header().Get(actions.LastLedgerHeaderName))
 
 		w = dbPathsClient.Get(uri + "?" + withSourceAccount.Encode())
 		tt.Assert.Equal(http.StatusOK, w.Code)
 		dbSourceAccountResponse := []horizon.Path{}
 		tt.UnmarshalPage(w.Body, &dbSourceAccountResponse)
+		tt.Assert.Equal("", w.Header().Get(actions.LastLedgerHeaderName))
 
 		tt.Assert.True(len(inMemorySourceAccountResponse) > 0)
 		tt.Assert.Equal(inMemorySourceAccountResponse, dbSourceAccountResponse)
@@ -237,11 +274,13 @@ func TestPathActionsInMemoryFinder(t *testing.T) {
 		tt.Assert.Equal(http.StatusOK, w.Code)
 		inMemorySourceAssetsResponse := []horizon.Path{}
 		tt.UnmarshalPage(w.Body, &inMemorySourceAssetsResponse)
+		tt.Assert.Equal("2", w.Header().Get(actions.LastLedgerHeaderName))
 
 		w = dbPathsClient.Get(uri + "?" + withSourceAccount.Encode())
 		tt.Assert.Equal(http.StatusOK, w.Code)
 		dbSourceAssetsResponse := []horizon.Path{}
 		tt.UnmarshalPage(w.Body, &dbSourceAssetsResponse)
+		tt.Assert.Equal("", w.Header().Get(actions.LastLedgerHeaderName))
 
 		tt.Assert.Equal(inMemorySourceAssetsResponse, dbSourceAssetsResponse)
 		tt.Assert.Equal(inMemorySourceAssetsResponse, inMemorySourceAccountResponse)
@@ -253,21 +292,18 @@ func TestPathActionsEmptySourceAcount(t *testing.T) {
 	defer tt.Finish()
 	orderBookGraph := orderbook.NewOrderBookGraph()
 	assertions := &Assertions{tt.Assert}
-	inMemoryPathsClient := pathFindingClient(
+	inMemoryPathsClient := inMemoryPathFindingClient(
 		tt,
-		simplepath.NewInMemoryFinder(orderBookGraph),
+		orderBookGraph,
 		3,
 	)
-	dbPathsClient := pathFindingClient(
+	dbPathsClient := dbPathFindingClient(
 		tt,
-		&simplepath.Finder{
-			Q: &core.Q{tt.CoreSession()},
-		},
 		3,
 	)
 
-	loadOffers(tt, orderBookGraph, "GA2NC4ZOXMXLVQAQQ5IQKJX47M3PKBQV2N5UV5Z4OXLQJ3CKMBA2O2YL")
-	loadOffers(tt, orderBookGraph, "GDSBCQO34HWPGUGQSP3QBFEXVTSR2PW46UIGTHVWGWJGQKH3AFNHXHXN")
+	loadOffers(tt, orderBookGraph, "GA2NC4ZOXMXLVQAQQ5IQKJX47M3PKBQV2N5UV5Z4OXLQJ3CKMBA2O2YL", 1)
+	loadOffers(tt, orderBookGraph, "GDSBCQO34HWPGUGQSP3QBFEXVTSR2PW46UIGTHVWGWJGQKH3AFNHXHXN", 2)
 
 	var q = make(url.Values)
 
@@ -294,12 +330,14 @@ func TestPathActionsEmptySourceAcount(t *testing.T) {
 		inMemoryResponse := []horizon.Path{}
 		tt.UnmarshalPage(w.Body, &inMemoryResponse)
 		assertions.Empty(inMemoryResponse)
+		tt.Assert.Equal("", w.Header().Get(actions.LastLedgerHeaderName))
 
 		w = dbPathsClient.Get(uri + "?" + q.Encode())
 		assertions.Equal(http.StatusOK, w.Code)
 		dbResponse := []horizon.Path{}
 		tt.UnmarshalPage(w.Body, &dbResponse)
 		assertions.Empty(dbResponse)
+		tt.Assert.Equal("", w.Header().Get(actions.LastLedgerHeaderName))
 	}
 }
 
@@ -308,9 +346,9 @@ func TestPathActionsSourceAssetsValidation(t *testing.T) {
 	defer tt.Finish()
 	assertions := &Assertions{tt.Assert}
 	orderBookGraph := orderbook.NewOrderBookGraph()
-	rh := pathFindingClient(
+	rh := inMemoryPathFindingClient(
 		tt,
-		simplepath.NewInMemoryFinder(orderBookGraph),
+		orderBookGraph,
 		3,
 	)
 
@@ -376,6 +414,7 @@ func TestPathActionsSourceAssetsValidation(t *testing.T) {
 			w := rh.Get("/paths/strict-receive?" + testCase.q.Encode())
 			assertions.Equal(testCase.expectedProblem.Status, w.Code)
 			assertions.Problem(w.Body, testCase.expectedProblem)
+			assertions.Equal("", w.Header().Get(actions.LastLedgerHeaderName))
 		})
 	}
 }
@@ -385,9 +424,9 @@ func TestPathActionsDestinationAssetsValidation(t *testing.T) {
 	defer tt.Finish()
 	assertions := &Assertions{tt.Assert}
 	orderBookGraph := orderbook.NewOrderBookGraph()
-	rh := pathFindingClient(
+	rh := inMemoryPathFindingClient(
 		tt,
-		simplepath.NewInMemoryFinder(orderBookGraph),
+		orderBookGraph,
 		3,
 	)
 
@@ -457,6 +496,7 @@ func TestPathActionsDestinationAssetsValidation(t *testing.T) {
 			w := rh.Get("/paths/strict-send?" + testCase.q.Encode())
 			assertions.Equal(testCase.expectedProblem.Status, w.Code)
 			assertions.Problem(w.Body, testCase.expectedProblem)
+			assertions.Equal("", w.Header().Get(actions.LastLedgerHeaderName))
 		})
 	}
 }
@@ -472,14 +512,14 @@ func TestPathActionsStrictSend(t *testing.T) {
 	destinationAssets, _, err := coreQ.AssetsForAddress(destinationAccount)
 	tt.Assert.NoError(err)
 
-	rh := pathFindingClient(
+	rh := inMemoryPathFindingClient(
 		tt,
-		simplepath.NewInMemoryFinder(orderBookGraph),
+		orderBookGraph,
 		len(destinationAssets),
 	)
 
-	loadOffers(tt, orderBookGraph, "GA2NC4ZOXMXLVQAQQ5IQKJX47M3PKBQV2N5UV5Z4OXLQJ3CKMBA2O2YL")
-	loadOffers(tt, orderBookGraph, "GDSBCQO34HWPGUGQSP3QBFEXVTSR2PW46UIGTHVWGWJGQKH3AFNHXHXN")
+	loadOffers(tt, orderBookGraph, "GA2NC4ZOXMXLVQAQQ5IQKJX47M3PKBQV2N5UV5Z4OXLQJ3CKMBA2O2YL", 1)
+	loadOffers(tt, orderBookGraph, "GDSBCQO34HWPGUGQSP3QBFEXVTSR2PW46UIGTHVWGWJGQKH3AFNHXHXN", 2)
 
 	var q = make(url.Values)
 
@@ -500,6 +540,7 @@ func TestPathActionsStrictSend(t *testing.T) {
 	accountResponse := []horizon.Path{}
 	tt.UnmarshalPage(w.Body, &accountResponse)
 	assertions.Len(accountResponse, 12)
+	assertions.Equal("2", w.Header().Get(actions.LastLedgerHeaderName))
 
 	for i, path := range accountResponse {
 		assertions.Equal(path.SourceAssetCode, "USD")
@@ -535,6 +576,7 @@ func TestPathActionsStrictSend(t *testing.T) {
 	tt.UnmarshalPage(w.Body, &assetListResponse)
 	assertions.Len(assetListResponse, 12)
 	tt.Assert.Equal(accountResponse, assetListResponse)
+	assertions.Equal("2", w.Header().Get(actions.LastLedgerHeaderName))
 }
 
 func assetsToURLParam(xdrAssets []xdr.Asset) string {