+++ services/horizon/internal/middleware_test.go
@@ -1,10 +1,17 @@
 package horizon
 
 import (
+	"net/http"
+	"net/http/httptest"
 	"strconv"
 	"testing"
 
+	"github.com/stellar/go/services/horizon/internal/actions"
+	horizonContext "github.com/stellar/go/services/horizon/internal/context"
+	"github.com/stellar/go/services/horizon/internal/db2/history"
+	hProblem "github.com/stellar/go/services/horizon/internal/render/problem"
 	"github.com/stellar/go/services/horizon/internal/test"
+	"github.com/stellar/go/support/db"
 	"github.com/stellar/throttled"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/suite"
@@ -141,4 +148,64 @@ func TestRateLimit_Redis(t *testing.T) {
 
 	w = rh.Get("/", test.RequestHelperRemoteAddr("127.0.0.2"))
 	assert.Equal(t, 200, w.Code)
+}
+
+func TestRequiresExperimentalIngestion(t *testing.T) {
+	tt := test.Start(t)
+	defer tt.Finish()
+
+	test.ResetHorizonDB(t, tt.HorizonDB)
+	q := &history.Q{tt.HorizonSession()}
+	app := &App{
+		historyQ: q,
+	}
+
+	request, err := http.NewRequest("GET", "http://localhost", nil)
+	if err != nil {
+		tt.Assert.NoError(err)
+	}
+
+	if q.GetTx() != nil {
+		t.Fatal("unexpected transaction in history.Q")
+	}
+
+	endpoint := func(w http.ResponseWriter, r *http.Request) {
+		session := r.Context().Value(&horizonContext.SessionContextKey).(*db.Session)
+		if session.GetTx() == nil {
+			t.Fatal("expected transaction to be in session")
+		}
+		w.WriteHeader(http.StatusOK)
+	}
+	appMiddleWare := appContextMiddleware(app)
+	handler := appMiddleWare(requiresExperimentalIngestion(http.HandlerFunc(endpoint)))
+
+	// requiresExperimentalIngestion responds with 404 if experimental ingestion is not enabled
+	app.config.EnableExperimentalIngestion = false
+	w := httptest.NewRecorder()
+	handler.ServeHTTP(w, request)
+	tt.Assert.Equal(http.StatusNotFound, w.Code)
+
+	app.config.EnableExperimentalIngestion = true
+	// requiresExperimentalIngestion responds with hProblem.StillIngesting
+	// if the last ingested ledger is 0
+	tt.Assert.NoError(q.UpdateLastLedgerExpIngest(0))
+	w = httptest.NewRecorder()
+	handler.ServeHTTP(w, request)
+	tt.Assert.Equal(hProblem.StillIngesting.Status, w.Code)
+
+	app.config.EnableExperimentalIngestion = true
+	tt.Assert.NoError(q.UpdateLastLedgerExpIngest(1))
+	tt.Assert.NoError(q.UpdateExpStateInvalid(true))
+	// requiresExperimentalIngestion responds with 500 if q.GetExpStateInvalid returns true
+	w = httptest.NewRecorder()
+	handler.ServeHTTP(w, request)
+	tt.Assert.Equal(http.StatusInternalServerError, w.Code)
+
+	app.config.EnableExperimentalIngestion = true
+	tt.Assert.NoError(q.UpdateLastLedgerExpIngest(3))
+	tt.Assert.NoError(q.UpdateExpStateInvalid(false))
+	w = httptest.NewRecorder()
+	handler.ServeHTTP(w, request)
+	tt.Assert.Equal(http.StatusOK, w.Code)
+	tt.Assert.Equal(w.Header().Get(actions.LastLedgerHeaderName), "3")
 }